# Conflicts:
#	calc_main/calc_main.go

+++ calc/calc_add.go
@@ -1,6 +1,6 @@
 package calc
 
 // Add takes two numbers and adds them together
-func Add(a int, b int) int {
+func Add(a float64, b float64) float64 {
     return a + b;
 }
