+++ calc/calc_subtract.go
@@ -1,6 +1,6 @@
 package calc
 
 //Subtract does
-func Subtract(a int, b int) int {
+func Subtract(a float64, b float64) float64 {
     return a-b;
 }
